requirements.txt
Django>=3.2.3,<3.3
gunicorn==20.0.4
djangorestframework